+++ include/pan_tilt/pan_tilt_node.h
@@ -34,6 +34,7 @@ private:
     int index0_tilt_trim_;
     int index1_pan_trim_;
     int index1_tilt_trim_;
+    bool first_reconfigure_;
     
     // Used to store last position (not including trim value)
     servo_msgs::pan_tilt index0_pan_tilt_;